Convert back to floats from doubles
To manage the previous issues with shadow acne, increase the margin for
the over and under points when computing shadow, reflection, and
transmission.

+++ src/camera.zig
@@ -16,18 +16,18 @@ pub const Camera = @This();
 
 horizontal_size: u32,
 vertical_size: u32,
-field_of_view: f64,
-half_width: f64,
-half_height: f64,
-pixel_size: f64,
+field_of_view: f32,
+half_width: f32,
+half_height: f32,
+pixel_size: f32,
 transform: mat.Matrix,
 
 pub fn camera(
-    horizontal_size: u32, vertical_size: u32, field_of_view: f64,
+    horizontal_size: u32, vertical_size: u32, field_of_view: f32,
     from: tup.Point, to: tup.Point, up: tup.Vector,
 ) Camera {
-    const h = @intToFloat(f64, horizontal_size);
-    const v = @intToFloat(f64, vertical_size);
+    const h = @intToFloat(f32, horizontal_size);
+    const v = @intToFloat(f32, vertical_size);
 
     var c: Camera = undefined;
 
@@ -68,8 +68,8 @@ fn transformCameraPerspective(from: tup.Point, to: tup.Point, up: tup.Vector) ma
 
 
 pub fn rayForPixel(c: *const Camera, px: u32, py: u32) ray.Ray {
-    const x_offset = (@intToFloat(f64, px) + 0.5) * c.pixel_size;
-    const y_offset = (@intToFloat(f64, py) + 0.5) * c.pixel_size;
+    const x_offset = (@intToFloat(f32, px) + 0.5) * c.pixel_size;
+    const y_offset = (@intToFloat(f32, py) + 0.5) * c.pixel_size;
 
     const world_x = c.half_width - x_offset;
     const world_y = c.half_height - y_offset;
